pyproject.toml
[project]
name = "constructbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiobotocore==2.15.1",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt==3.2.0",
    "fastapi-versionizer>=4.0.2",
    "fastapi[all]>=0.117.1",
    "geoalchemy2>=0.18.0",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "pre-commit>=4.3.0",
    "python-jose>=3.5.0",
    "ruff>=0.13.1",
    "shapely>=2.1.1",
]

[tool.ruff]
fix = true
exclude = ["docs", "**/migrations"]
target-version = "py312"
lint.select = [
  "E",
  "F",
  "I",
  "UP",
  "YTT",
  "S",
  "T20",
  "PIE",
  "PT",
  "Q",
  "RUF",
]
lint.ignore = ["RUF100", "RUF003", "RUF001", "RUF002", "S106", "S105"]
line-length = 120
indent-width = 4

[tool.ruff.format]
quote-style = "double"  # или 'single'
docstring-code-format = true
docstring-code-line-length = 88

[tool.ruff.lint.per-file-ignores]
"app/tests/**" = ["S101", "E501"]
